Renamed some source files for clarity.
