+++ install/merc.sql
@@ -29,10 +29,6 @@ INSERT INTO area VALUES (29,1,'Brannington Town',1,1232009988,9943,1,27613,160,1
 INSERT INTO area VALUES (30,1,'Clan Spawners',0,1232009989,9995,1,27614,0,12279808,0);
 INSERT INTO area VALUES (36,1,'Caligar',0,1232009991,9896,1,27616,0,14950400,0);
 INSERT INTO area VALUES (33,1,'Long Tunnel',0,1232009991,9953,1,27615,0,21962752,0);
-INSERT INTO area VALUES (1,7,'The Town of Cameron 7',0,1232009960,9953,1,27585,0,17211392,0);
-INSERT INTO area VALUES (1,13,'The Town of Cameron 13',0,1232009961,9952,1,27586,0,17256448,0);
-INSERT INTO area VALUES (1,20,'The Town of Cameron 20',2,1232009962,9940,1,27587,843,18137088,0);
-INSERT INTO area VALUES (4,14,'Pentagram Quest 14',2,1232009966,9898,1,27591,1038,18497536,0);
 INSERT INTO area VALUES (37,1,'Arkhata',0,1232009993,9836,1,27617,0,17973248,0);
 
 INSERT INTO vars VALUES ('cur_on',0);
