+++ src/aston.c
@@ -22,9 +22,9 @@
 #include "libload.h"
 #include "quest_exp.h"
 #include "item_id.h"
-#include "area3.h"
+#include "aston.h"
 #include "los.h"
-#include "staffer_ppd.h"
+#include "misc2_ppd.h"
 #include "questlog.h"
 #include "statistics.h"
 #include "player.h"
