+++ src/brannington.c
@@ -25,7 +25,7 @@
 #include "libload.h"
 #include "sector.h"
 #include "player_driver.h"
-#include "staffer_ppd.h"
+#include "misc2_ppd.h"
 #include "questlog.h"
 #include "arkhata.h"
