+++ src/cameron.h
