+++ src/questlog.c
@@ -11,14 +11,14 @@
 #include "tool.h"
 #include "depot.h"
 #include "drvlib.h"
-#include "area1.h"
-#include "area3.h"
+#include "cameron.h"
+#include "aston.h"
 #include "database.h"
 #include "questlog.h"
 #include "btrace.h"
 #include "player.h"
-#include "staffer_ppd.h"
-#include "two_ppd.h"
+#include "misc2_ppd.h"
+#include "exkordon.h"
 #include "container.h"
 #include "chat.h"
